fix(flash/cards): fix card texture mix-up bug and ECS architecture issues
BUG FIX (card texture cross-contamination):
- destroySceneEntity() now clears cardDataStore[eid] and shaderStyleStore[eid]
  so that when bitECS reuses an eid, the new entity does not inherit stale data
- CardRenderer destroy callbacks now set glStateStore[eid] = undefined and
  gpuStateStore[eid] = undefined after deleting/releasing the GPU resource
  (previously the stale GLStateData with a deleted WebGL texture remained,
  causing the new entity occupying the same eid to render the wrong card face)

Refactor:
- Extract zoomToLodScale() to utils/glUtils.ts; import in both glTextureSystem
  and gpuTextureSystem (removes 3 duplicate LOD implementations)
- Merge RenderConfig + GPURenderConfig into a single interface in constants.ts;
  both render systems re-export it for backward compatibility
- Remove implicit bootstrapCards() module-side-effect from textureRenderer.ts;
  call it explicitly in CardRenderer constructor instead
- Fix registerBulk() skipping change listener notification
